hibernate: Lock the feed when adding entries
Otherwise, there's a race condition where newer entries can get older
timestamps: if server A starts a transaction, gets a timestamp, but then
gets blocked, and server B then starts a transaction, gets a higher
timstamp, and commits, then when server A resumes, its entry will be
written at a later time then the entry from server B, but with a lower
timestamp. A consumer that read the entry from server B and used it as a
marker would then not get the entry from server A. Locking the feed
before getting the timestamp and holding it until the commit solves this
issue, at the cost of serializing all writes to a feed.