hibernate: Mark timesetamps unique
Pagination doesn't work well when a marker has the same timestamp as
other entries. If one producer writes an entry, a consumer reads it, and
then another producer writes a new entry, and this gets the same
timestamp as the first entry, then it's not clear whether the consumer
should get this second entry or not when using the first entry as
marker. If entries with the same timestamp as the marker should be
included, then consumers may get entries multiple times if they share
timestamp with the marker. If entries with the same timestamp as the
marker should not be included, then consumers may miss entries.

Using secondary sorting on ID also doesn't work, since it then is
possible for a new entry that happen to get a lower-sorting ID to be
considered older than an existing entry with a higher-sorting ID.

This commit prevents these error conditions by forbidding sharing of
timestamps between entries in the same feed. That way, a producer that
happens to get the same timestamp as an earlier entry will have to
retry, giving it a new timestamp. This ensures that consumers get each
entry exactly once.

Due to the high resolution time stamps and the lock on the feed,
collisions should in practice never happen.