hibernate: Use database timestamps for entries
If application servers have clocks that are not in sync, then it is
possible for timestamps to not be increasing. If there are two
application servers with out-of-sync clocks, where the clock of server A
is a bit ahead of the clock of server B, then it's possible that an
entry written by A gets a higher timestamp than en entry written at a
later point by server B. If a consumer read the entry written by server
A, and then used it as a marker, then it would not receive the entry
written by server B.

This change makes all application servers ask the database server for
the clock, similar to how it is done in the Postgres adapter.