test: Programs for verifying following by marker
MarkerRaceConditionTest posts multiple entries in parallel, while
following the feed using the marker mechanism. The code is written to
try to maximize the risk of having multiple entries being handled in
parallel by the server. If any entry is written with an older timestamp
than the latest, then that entry will be missed, and this will be
revealed by the program.